Trijsturis un aplis

+++ src/manPatikStradat/Trijsturis.java
@@ -4,7 +4,7 @@ public class Trijsturis {
 	public double mala1,mala2,mala3;
 	public Figuras punkts;
 
-	public Trijsturis(Figuras p, double m1, double m2, double m3) {
+	public Trijsturis(Figuras p, int m1, int m2, int m3) {
 		punkts = p;
 		mala1 = m1;
 		mala2 = m2;
